test/e2e: migrate serving-cert-annotation test for OTE compatibility
Migrate the serving-cert-annotation test to support dual execution:
- Standard Go test: go test -v ./test/e2e -run TestE2E
- OTE framework: service-ca-operator-tests-ext run-suite

Changes:
- Add test/e2e/serving_cert.go with Ginkgo wrapper and shared test logic
  using testing.TB interface for dual compatibility
- Update cmd/service-ca-operator-tests-ext/main.go to register the new
  test suite with OTE framework
- Simplify test/e2e/e2e_test.go inline test to call shared function
- Move pollInterval/pollTimeout constants to serving_cert.go for package
  builds
- Use clientcmd.BuildConfigFromFlags to avoid stdout output interfering
  with OTE JSON parsing

The test implementation is in serving_cert.go for dual compatibility.
Once OTE is stable, the inline test in e2e_test.go will be removed.

Test tags: [sig-service-ca], [Operator], [Serial]
